requirements.txt
flask==2.0.1
flask-login==0.5.0
psycopg2-binary==2.9.9
bcrypt==3.2.0
gunicorn==21.2.0
requests==2.28.1